phpcs: Remove comment regarding silence (at-sign) operator in tests
This comment was added in 355a6cd4c3045202 and quickly became outdated.
While it should still mostly be correct, it is no longer "policy" to use
the at-ease library. The "@" operator is now explicitly permitted where
necessary and is now recommended even outside tests (T253461). As such,
it no longer makes sense to keep this comment around.

Change-Id: I07aa84d46be8c7e5a1c8a3249d382cbbfef80267