Allow jquery.tipsy on mobile
This module deprecated, but it doesn't mean we should strictly not
load this on mobile. We do the same for jquery.ui.

Not targetting mobile means we can't migrate certain extensions
e.g. Wikibase off of the targets system.

This should be safe as nothing directly calls addModule on
this library.

Bug: T323542
Change-Id: I3e06d1fe0b82ac9a8d926702afcf2058662465b6