Fix iw-widget/footer overlap
Icea9a103e62cd353d14ec230bec44eff311f54e4 refactored
positioning of a bunch of elements and went with
absolutely positioning the interwiki widget on the
right hand side. Rendering this out-of-container has
the unforeseen consequence of the interwiki widget
overlapping with the footer in cases where it is
longer than the search results on the left.

This takes a different approachs: the nodes on the
left (`.mw-search-results-container` and
`.mw-search-results-info`) are now floated left,
leaving the space on the right available for the
interwiki widget, now `display: inline-block`, to
fill. Because it's no longer actively positioned to
the right, we must, however, make sure that that
widget now gets the margin-left to separate it from
the main search results.

This positioning is a little more prone to failure
than what we had: positioning has to be pixel perfect
or the widget will not fit in the right-hand-side
gap left by the 2 left-floated nodes. And, as said
earlier, the whitespace in between them must now be
explicit. But since recent refactorings, the rest of
the CSS is now in a much better state where we can
actually rely on things consistently being where we
want them to be, without skin overrides or other
situational (e.g. interwiki widget presence)
rendering discrepancies.

Note that I also removed a `margin-top: 24px`.
The search form (`.mw-search-form-wrapper`) already
has a `margin-bottom: 24px` and that ended up being
collapsed with said margin-top. That is now no longer
the case with that parent node being floated, and it
would result in a double 24px margin.

Bug: T324428
Change-Id: I75d6d2195eb4fe2edf47d377f85fb67914937dd5