Allow uses of HTMLForm to specify a minlength for text based inputs
Allow HTMLForm uses to specify a minlength attribute for text based
inputs. This includes the SelectOrOther field but does not include
the SelectAndOther field as further work would be needed to
achieve that because a custom javascript function to validate that
the length of the selected item and the other field is above the
minimum required.

Bug: T321427
Change-Id: I30d54f44d2a164615bfd248322eaa4f12270c9b1