linker: Fix broken specialLink() when combining subpage with querystring
Follows-up 3255f37af1 (Ibcd2c5365a1f581e), which introduced support
for subpage slashes and querystrings. Due to the order of operations,
it split off the subpage (including query) before the check for the
query, which meant it was actually encoded as part of the subpage.

The original unit test did expose this issue but the bug was part of
the expected value.

For example:
<https://en.wikipedia.org/wiki/Special:Contributions/JohnDoe%3FtopOnly%3D1>
vs
<https://en.wikipedia.org/wiki/Special:Contributions/JohnDoe?topOnly=1>
of which only the latter is identical to
<https://en.wikipedia.org/?title=Special:Contributions/JohnDoe&topOnly=1>

While at it, also improve readability and future-proof stability by
making the conditional check and the substring operation share the
same result directly instead of computing it three times. This avoids
future refacting causing the three lines to go out of sync in a way
that e.g. the subsequent strpos operations might find false which
would lead to an unpredictable outcome. Phan isn't able to catch this
yet, but probably will in the future, namely that strpos() can return
false which isn't a valid param to substr().

Change-Id: Iec00a6ecbf7abf57b8944e32a6256ac9db4a653b