Title.js: Rename Title.getName[Text]
These methods have very misleading names which easily leads to
confusion with getMain[Text] which is the right method 99% of the
time. Rename them to clearly indicate they are filename-related
(it would be nicer to make them a no-op or error on non-file titles
but it would cause a B/C break which is not worth the effort) and
deprecate the old methods.

Bug: T324741
Change-Id: Ibd42236eff9e73e392c0686767c388dbc7bb249c