Don't enable logging linter data in the /lint/ endpoint
We only want to log full page parses of the latest revision and that's
already covered by the wt2html endpoint and changeprop should ensure it
always happens.  In the future, we may want to restore this when
exploring ways to avoid the performance cost of doing the linting during
the canonical parse, but that's T325031.

Follow up to I1f69498ef759f7a82ad8ad9002d7212636e92ffe

Bug: T325031
Change-Id: Icefb46b416629c2714a7d4f282cd55cbca271323