Enable several modules on mobile site
These modules are often loaded on the page and currently silently removed.
Looking through the code, there is nothing problematic with running
any of them on mobile. The code relates to workflow and special
page (or logged in traffic preferences)
so shouldn't impact mobile first paint on article views.

The following modules are added:
* mediawiki.page.image.pagination
* mediawiki.action.edit.collapsibleFooter
* mediawiki.filewarning
* mediawiki.userSuggest
* mediawiki.rcfilters.highlightCircles.seenunseen.styles
* mediawiki.special.watchlist
* mediawiki.special.userrights
* mediawiki.misc-authed-pref
* mediawiki.action.view.postEdit

Bug: T324723
Change-Id: Ie0a49e5f37ac5e47066ce74d7e618d933cb30809