rcfilters: Prevent tagfilter param from being set to 'all'
On a wiki with very few tags, it's easy to select all tags. Because the
tags group is a string_options group, this results in a query string
with &tagfilter=all, which is incorrect and returns no results. (In the
unlikely event that a tag named 'all' exists, it will return confusing
results instead and cause other bugs.)

Fix this by adding a 'supportsAll' property to string_options groups,
which defaults to true and is only set to false for the tagfilter and
namespace groups. When supportsAll is false, 'all' is not generated or
interpreted as a magic value.

Bug: T325255
Change-Id: I892c9e2893de5b5463da76696db01e556d3e187f