actions: Optimize handling of notification timestamp on action=history
HistoryPager::formatRow is called for each row of the database result,
when the config is true the WatchlistManager is called to get the
timestamp of the last notification for a user and title, but this is for
each row the same user and the same title, use a class property to store
the information in the pager instead and not relay on the cache in the
WatchlistManager. The value cannot change between the calls.

Collapse the now simple call to historyLine()

Change-Id: I8807c32469b838a93020a86b6d192a9ed290cba0