rdbms: various fixes to LoadBalancer::reconfigure
Use server names to handle the case where server indexes get
shifted around due to the depooling or provisioning of a server.

Previously, loads could be assigned to a wrong server, assigned
to a server that was depooled, or assigned to a server that is
to new to appear in the current "servers" array. This would mean
that getReaderIndex() or getLagTimes() could trigger exceptions.

Only change loads for depooled servers. Update the loads of other
servers to the current loads only makes sense in the context of
using all the current servers. This does not apply to reconfigure()
since it does not see newly pooled servers (for simplicity).

Handle servers depooled only from custom query groups.

Bug: T322156
Change-Id: I9f710aa32f5d5b74796bb80a8426a5f653b8e4d3