tests: Do not set User::mBlock in secondary auth provider test
The user block was set directly on the User objects,
instead change the enviroment to load the correct block.

This is done by inserting the block into the database via
DatabaseBlockStore, that is enought for user blocks.
For ip blocks also the connection between the blocked ip and the testing
user has to be prepared, that is done via global state, similiar to a
user session.

Not setting User class internals helps on deprecation and remove of
class properties

Change-Id: I4ceec2c1a998e962648669c43fb8cf812fab8432