rcfilters: Tweak formatting of inverted filters
It seems to me that the "not" should apply to the whole phrase
"#Foo" / ":Foo" ("tagged with Foo" / "in namespace Foo"),
rather than be inserted in the middle of it.

I think this will also encourage better translations in languages
where the "not" goes at the end of the phrase. Currently they do
things with the "#" / ":" marks that don't really make sense to me.

Change-Id: Ie3fa4e37e5f9df58fc38c097d5e4549c47252128