phpunit: Restore CI log of "rdbms" and "wfDebug" channels
In CI, when a test fails, all messages received by any debug logger,
channel at any severity, are meant to be captured and printed after
the assertion details are printed.

This did not work for the "rdbms" channel, which effectively never had
any of its messages printed. This is because the LBFactory service is
created only once during the first "group Database" test and then
globally kept, regardless of our services container resets, and thus
never picks up LogCapturingSpi.

From a service container perspective, this is intentional and makes
sense, because we need to manage the temporary database through this
not just between separate TestCase subclasses, but even within one
TestCase subclass between test method (e.g. addDataOnce).

These Logger objects are created before or during the first (and likely
passing) test, and after that we lose access to them in
MediaWikiLoggerPHPUnitExtension, but they remain held inside the Rdbms
service.

Fix this by refactoring LogCapturingSpi to exist continuously
instead of creating a new one each test. After a test is finished,
we clear the log of the persistent instance.

MediaWikiIntegrationTestCaseTest has a test case that asserts
referential identity mismatch. I don't know why it cared about this.
It was added I4dee53176f (406b844a42) as part of a larger refactor
without explanation.

Change-Id: I97f00a8db8b77ad36eabe0b38db986d775961a23