Include missing message parameter in message
Message parameter was visible to a user when moving a page. This is
undesirable as the console gives incomplete feedback. This patch
solves that by including the 'old name' of the page to be moved.

Bug: T322583
Change-Id: Ie155d5c3d588702ad169fba2563b3ee9f50dd292
(cherry picked from commit f79d0469e42787e566dde9b794a0d5f283b4e1f0)