maintenance/mysql.php: Add --raw-host parameter
This is to ease connecting to ParserCache and MainStash DB hosts, which
are configured directly with SqlBagOStuff instead of via $wgLBFactory,
thus they are considered unknown to the way this script worked before.

This new option can also be used to connect with regular core DB hosts
that e.g. are not yet pooled or are depooled for some reason.

Change-Id: I76f362d10818ec2ea4d8f3fbcb367835928be7de