Galleries: Make gallery code more resilient to HTML transformations
The gallery code should be more resilient to HTML transformations
that can occur inside mobile app and MobileFrontend (or gadgets).

Data attributes are generally left untouched, so provide this
information as part of the output of the LI element so that the
gallery code can make use of it where possible.

This resolves the problem of lazy loaded images in the mobile site
meaning the hook added for MobileFrontend can also be removed.

Follow up to I58e18924c02fe12baba820f25b7a6f0c437a206b

Bug: T194887
Change-Id: I186bb0dbbea84d4dbb1d7b55278c69b2b46603de