Parser: Add mw-selflink-fragment class to links to hash fragments in the same page
It should be cheap for JavaScript modules to identify all links in the page
which point to hash fragments within the current article. These should not be
styled the same as self links given the existing parser text.

This will be used by Popups to disable previews on links that link back
to the same page.

Bug: T198652
Change-Id: Ic6cca5023c32291032cf10c1ece9c70ddd0f40c2