Galleries: Set anchor height as well as img height
In responsive skins, image height is often set to auto.
If we set the parent element height as well this allows those
skins to use a skin style to correct the heights for gallery
images as done in I3904bb158eebe96983ce3318e1e39a6f056624fd

Additional change:
Drop "px" from width and height declarations. These are not necessary.

Bug: T194887
Change-Id: I5c372f8b1bbc44af9c5855baa7ee825ff42fbc96