Make ChangeTag::modifyDisplayQuery more reliable if multiple joins to changetags
This adds a table alias for the added SQL conditions, so if there
is a separate join to the change_tag table, it won't step all over it.

This is meant for compatibility with 69df5d906dc8 when both a change tag
and the rendered revision filter is active at the same time.

Bug: T325872
Change-Id: I56a349028471c18003291cbed411b1a0535630bc