Add support for the Igala language (igl)
There are several variants for the spelling of
the autonym, with and without diacritics.
After consultation with User:AgnesAbah,
who contributed articles to Wikimedia Incubator
and wrote all the translatewiki translations,
I'm defining it using a simple string, which
is identical to English and without any diacritics.
After a community forms, this can be reconsidered.

Bug: T326526
Change-Id: I7b97f1f7be2c6e00d5f2a4254baf2262d08465c0