MaintenanceRunner: Print error if Maintenance initialization fails
This is the replacement for the check I added in b0647c30bc30874c and
removed in I8032ab18cb26d92e:

If an Exception or Error occurs while the Maintenance object is being
constructed, or during the call to setName(), print an error message
with a reminder that scripts need to call parent::__construct(), and
also the exception details normally displayed and/or logged by PHP's
default error handler, then exit with a nonzero status.

I think this should work well enough. More precise targeting of the
reminder message likely would involve some sort of ugly hack and is
probably not necessary. It simply wouldn't be worth it, given that
the defective constructor likely would be in the stack trace anyway.

Follows-up 5bfa31187b266f83 and a9f47fd5ad4ea668.

Change-Id: I5a4ee2b23e99e195469cf258ae3833d8e18bc7b8