PoolCounterRedis: Fix wrong cast, locks weren't being released
The slotTime is generated with the php function microtime( true ). The
first argument causes microtime to return a float value instead of an
integer value.

slotTime is used for CAS-style check. The value was being sent as float
for locking, but as integer when releasing. Since both values didn't
match, the release was ignored.

Follow up to Icb2de32107f43817acc45fe296fb77acf65c1786

Bug: T326886
Change-Id: I47383f9f9441a62fa64827ec50b1e6984c6804cb