MaintenanceRunner: improve error reporting
This improves the error message produced by run.php when a script is not
found or cannot be executed.

In particular, this fixes the isse that run.php would say it can't find
a script even if the script was indeed found, but was a "plain" php
script that doesn't set $maintClass.

Change-Id: Ia45c7fdaf3f7f631f8267e7e645f13fe672955f4