specials: Fix links to revisions in diff on Special:Undelete
The existing check for revisions without id and title does not longer
work, rows in the archive table have a revision id as well (since 1.31
the field ar_rev_id is not null), the title was passed in ba99527 and
broke the check as well.

In the diff mode there was no link to the content of the deleted
revision from the diff itself.
If the previsous revision is not deleted, the link works as expected.
Links to the deleted revision gives error "The revision #n of the page
named "p" does not exist.", deleted revisions only viewable via
Special:Undelete

Remove the fixmes as that is no longer relevant with RevisionRecord and
after ba99527, it was true when added (90ce609)

Bug: T44608
Change-Id: Ieacafcf99a2f3aa1dfe14b3704899ee2c218e0f0
(cherry picked from commit 1f4c055d9a196b31c388d463b9c89b9a942f032c)