Maintenance: find file by name in maintenance dir
This allows a file in the maintenance dir to be specified without the
path but with the .php extension. Previously, the file would be expected
to be found in the local directory if it had no path but a .php
extension.

This effectively makes it so that the maintenance directory is used as
the default location for maintenance script files, which can be
specified with or without .php extension. The extension is no longer
significant for how the file is located. Only paths that start with
"./" or "../" are treated as relative to the cwd.

Depends-On: Ic1ca796a52a0c5d7095e607fdc2e320b086229c7
Bug: T326800
Change-Id: Id334d2b3ccd4be1443d95eb40b028f25c47f94c0