API: Allow use of multivalue parameter documentation for templates
Allow the use of PARAM_HELP_MSG_PER_VALUE for PARAM_TYPE='string'
and PARAM_ISMULTI=true, which is the setting used for templated
parameters (PARAM_TEMPLATE_VARS).

It's common for templated parameters to take the variable parts
of the name from some fixed set, but they are required to use
PARAM_TYPE=string so it makes sense to allow the use of multi-
valued parameter documentation for that case.
The other option would have been to allow templated parameters
to be enums, which also makes sense and could be a complimentary
improvement; there are two reasons to prefer the current approach:

* Getting the list of allowed values might have a performance
  impact that we only want to incur when the (optional) parameter
  is used, while an enum requires the value list to be generated
  every time the API module is used. So there would be some
  value in using string type for templated parameters with a
  known set of documentable values even if enum were also allowed.
* Per-value documentation might be useful outside the template
  use case for non-restricted multi-valued strings: maybe all
  values are allowed but there are some values with special
  meanings that are worth documenting.

Change-Id: I53f9ae840c0a7eee76c4b57f95390b5045707efd