Special:ApiHelp: Move widths and floats in CSS to media query
Several widths in Special:ApiHelp are large and so don't play nicely with mobile. This change moves
most "wide" width items to a min-width (720px) media query, allowing the skin to supply the
"default" rules. Floating and related rules are also moved. Changes were tested in both the output
of Special:ApiHelp as well as its transcluded version.

Bug: T327220
Change-Id: I6f35967125fe00434a5aed49ab2210c6d590335c