ApiOptions: Give warning if the value is too long
I was thinking add validations to HTMLApiField and other field types
if any, but then I found one field type that hard to implement proper
length validation, which is used for the `email-blacklist` preference.

The API will not save options that exceed the limit after this patch,
as if the validation is done by respective html fields.

Bug: T326696
Change-Id: Ib6b04db33bc6a806a49a16a1797c191faf5fb59d