rdbms: improve the moving average method in LoadMonitor
Make the moving average ratio dynamic and generally much
higher. It scales up based on how far in the past the last
time the server was polled. This makes it better handle a
wide range of traffic (e.g. large vs small sites or web vs
CLI mode) and converge faster on recognizing a down server.

Bug: T265386
Change-Id: I155b42bb93a512b17e244b58d5c4cd55fa478886