rdbms: tweak the refresh probability in LoadMonitor
Use a power function rather than linear function to better handle
a wide range of possible traffic. Since the new probability curve
bows under the linear one, the chance of a refresh a few dozen ms
after caching is low, even when the access rate is high. It still
catches up to the old curve once the cache age is around the max.

Bug: T265386
Change-Id: I6acb3a133a5ac997e1e05a561cc8e61c3fb52431