Lists of images should use lazy loading
To improve on problems with 429s thumbor requests, we should
simply delay loading of images in lists with native lazy loading.
This will cause fewer requests to be queued initially on desktop
browsers, making it less likely to trigger rate limiting.

Enable for:
- Image query results
- Image pager
- File history list
- Search results

Not enabled for normal thumb and gallery transclusions (in part
because we still have mobile JS lazy loading)

Bug: T266155
Change-Id: I5a186780053d9e098101e8106b1294a8016b7a04