ParserTestRunner: set the user language to variant when present (take 2)
In production the user language is set to the variant code, not the base
language code, and the parser test runner should match that.  Similarly,
we don't need to explicitly ::setLanguage() in the RequestContext; let
it pick up the appropriate language (actually, variant) from ::setUser().

This exposes a bug in the way ToC conversion is done, which results in
double-conversion of the ToC title.  This will be fixed in
I321cd31dae64bbf845d53282e5d28a55bc4ec319.

Some of the variant tests use Pig Latin as the target variant.  But in
order to have good results from parser tests with Pig Latin as a
variant, it needs to be "supported" aka have at least one localized
message, so we can verify that the correct localization is being done.
Some tweaks were needed to LanguageNameUtils in order to ensure that
Pig Latin is supported if (and only if) the UsePigLatinVariant is
enabled in the site configuration.

This is a revision of Ia72c960b2c7914342eb4d5e3e63f2d6af14719ad, which
had to be reverted when it broke Parsoid CI: the ParserTestRunner
used $context->getLanguage() in some places (which is the *user*
language) where it ought to have used the *page* language.

Depends-On: I318c2fb8694c90efef07f1e2951c6d9aa6b3e82f
Change-Id: I0c9c9fec920f7cb028d935e552a8f11475a23ba7