ParserTestRunner: properly set the user language when a variant is present
In production the user language is set to the variant code, not the base
language code, and the parser test runner should match that.  Similarly,
we don't need to explicitly ::setLanguage() in the RequestContext; let
it pick up the appropriate language (actually, variant) from ::setUser().

This exposes a bug in the way ToC conversion is done, which results in
double-conversion of the ToC title.  This will be fixed in
I321cd31dae64bbf845d53282e5d28a55bc4ec319.

Some of the variant tests use Pig Latin as the target variant.  But in
order to have good results from parser tests with Pig Latin as a
variant, it needs to be "supported" aka have at least one localized
message, so we can verify that the correct localization is being done.
Some tweaks were needed to LanguageNameUtils in order to ensure that
Pig Latin is supported if (and only if) the UsePigLatinVariant is
enabled in the site configuration.

Depends-On: I318c2fb8694c90efef07f1e2951c6d9aa6b3e82f
Change-Id: Ia72c960b2c7914342eb4d5e3e63f2d6af14719ad