Language-convert Table of Contents at parse time
In 24949480eb22947298e6a82db06c55af0d262db2 (Oct 2021) injection of
the Table of Contents was moved from Parser to
ParserOutput::getText(); that is, from parse time to "postprocess text
possibly fetched from the cache" time.  Unfortunately, this meant that
language conversion wasn't done on the table of contents (!), for
either traditional skins or the vector-2022 skin.  This was fixed for
traditional skins by 059e62cde6cdf4a87dfa45a79191bed531045501 (Nov
2021), later amended by 0955046ca59aae705fee8b098e4cfef394647b5e (Mar
2022), which added explicit language conversion to the TOC injection
process in ParserOptions::getText().  This fix was still not complete,
however, since editor-defined custom language-conversion rules defined
in the article body were no longer available to the language converter
when conversion was done in ParserOutput::getText(); the ToC title was
also being double-converted.  Further, neither of these short-term
fixes addressed the output of ParserOutput::getSections() (now
ParserOutput::getTOCData()) which was used by vector-2022 to generate
the ToC in the sidebar and which remained entirely unconverted.

With 439656e01974e3cbc57902fb4ef33e71c1a386d5 (Jan 2023), we started
using the ::getSections()/::getTOCData() output for main article text
as well, but we kept the previous hack which post-converted the
generated HTML. This kept old skins at parity with the post-Oct-2021
status, but also didn't address the conversion issue for vector-2022.

The solution here is to perform language conversion on the ToC lines
at parse time along with the rest of the language conversion, and
store *converted* headings in TOCData.  This has a number of side
effects:

1. The ToC information array available via the action API
is now language converted.  This is *probably* what you wanted in the
first place, but could potentially be disruptive.

2. The ToC is consistently converted with the full set of
editor-defined custom conversion rules.  Before Oct 2021, the ToC was
converted using the set of custom conversion rules *active at the
point at which the ToC was inserted* (which was usually near the
beginning of the article).  When all conversion rules appear at the
very top of the article (best practice!):

 -{en:Foo; en-x-piglatin:Bar;}
 Lead section text
 == Introduction ==
 == Foo ==

There should be no difference before pre-Oct 2021 behavior and the
behavior after this patch: in both cases the rule defined in the
article body will be applied both to the heading and to the TOC, and
they will be consistent.  (After Oct 2021 and before this patch, Foo
would be converted in the heading but not in the table of contents.)

But in cases where conversion rules are defined after the
TOC insertion point, the section heading as it appears in the body
text could appear different from the section heading as it appears in
the ToC.  For example, if you defined a conversion rule just before
using a term in a heading:

 == Introduction ==
 -{en:Foo; en-x-piglatin:Bar;}-
 == Foo ==

Before Oct 2021, this rule would be applied to the heading, but not to
the TOC (because the TOC insertion point was before the rule
definition).  This would also be the behavior before this patch (since
rules defined in the article body are currently not applied at all).
After this patch, the rule will be applied to both the heading and the
TOC (because the rule application location is effectively "at the very
end of the article").  In the rare cases when rules are not defined in
glossaries at the top of the article, this type of usage (definition
immediately preceding first use) is expected to be the most common
and the behavior after this patch is more correct.

But alternatively, if you defined a conversion rule *after* using
the term in a heading:

 == Introduction ==
 == Foo ==
 -{en:Foo; en-x-piglatin:Bar;}-

Before Oct 2021, this rule wouldn't be applied to the heading *or* the
TOC.  Before this patch, this would also be the case (because rules
defined in the article body are not applied at all).  After this
patch, the rule will be applied to the ToC but not the heading, since
the application point for the TOC is effectively at the end of the
article.  This inconsistency is probably not desirable, but this case
is expected to be rare, and (assuming the editor intended 'Foo' to be
unconverted) the editor can work around the inconsistency by
explicitly protecting 'Foo' from conversion:

  == -{Foo}- ==
  -{en:Foo; en-x-piglatin:Bar;}-

And if the editor /intended/ Foo to be converted, the rule definition
should be moved earlier in the article.  Again, putting all rules at
the top of the article is the preferred style, and works better with
the glossary style used by the zhwiki community (see also
https://www.mediawiki.org/wiki/Requests_for_comment/Scoped_language_converter
).

Bug: T306862
Depends-On: I0c9c9fec920f7cb028d935e552a8f11475a23ba7
Change-Id: I321cd31dae64bbf845d53282e5d28a55bc4ec319